split trivial_op.cc
