--- paddle/cinn/hlir/framework/pir/trivial_op.h
@@ -1,43 +1,0 @@
-// Copyright (c) 2024 PaddlePaddle Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-#pragma once
-
-#include "paddle/cinn/hlir/dialect/operator/ir/manual_op.h"
-#include "paddle/cinn/hlir/framework/compile_error.h"
-#include "paddle/cinn/hlir/framework/pir/op_lowering_util.h"
-#include "paddle/cinn/hlir/framework/pir/utils.h"
-#include "paddle/cinn/hlir/op/external_api_registry.h"
-#include "paddle/cinn/hlir/pe/map_expr_to_ir.h"
-#include "paddle/cinn/ir/dim.h"
-#include "paddle/cinn/ir/group_schedule/base_group_scheduler.h"
-#include "paddle/cinn/ir/group_schedule/st_shape_group_scheduler.h"
-#include "paddle/cinn/ir/schedule/ir_schedule.h"
-#include "paddle/cinn/lang/placeholder.h"
-#include "paddle/cinn/optim/schedule_block_dce.h"
-#include "paddle/cinn/optim/transform_gpu_forloop.h"
-#include "paddle/common/ddim.h"
-#include "paddle/fluid/pir/dialect/operator/ir/op_type.h"
-#include "paddle/pir/include/dialect/control_flow/ir/cf_op.h"
-
-namespace cinn {
-namespace hlir {
-namespace framework {
-namespace pir {
-std::vector<ir::Expr> OperationFusion(
-    const std::vector<::pir::Operation*>& ops,
-    const std::vector<ir::Expr>& op_compute_bodies);
-}
-}  // namespace framework
-}  // namespace hlir
-}  // namespace cinn
