+++ paddle/cinn/hlir/framework/pir/trivial_op_impl.h
@@ -1,0 +1,209 @@
+// Copyright (c) 2024 PaddlePaddle Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+#pragma once
+
+#include <variant>
+
+#include "paddle/cinn/hlir/dialect/operator/ir/manual_op.h"
+#include "paddle/cinn/hlir/framework/compile_error.h"
+#include "paddle/cinn/hlir/framework/pir/op_lowering_util.h"
+#include "paddle/cinn/hlir/framework/pir/trivial_op_util.h"
+#include "paddle/cinn/hlir/framework/pir/utils.h"
+#include "paddle/cinn/hlir/op/external_api_registry.h"
+#include "paddle/cinn/hlir/pe/map_expr_to_ir.h"
+#include "paddle/cinn/ir/dim.h"
+#include "paddle/cinn/ir/group_schedule/base_group_scheduler.h"
+#include "paddle/cinn/ir/group_schedule/st_shape_group_scheduler.h"
+#include "paddle/cinn/ir/schedule/ir_schedule.h"
+#include "paddle/cinn/ir/schedule/ir_schedule_util.h"
+#include "paddle/cinn/lang/placeholder.h"
+#include "paddle/cinn/optim/schedule_block_dce.h"
+#include "paddle/cinn/optim/transform_gpu_forloop.h"
+#include "paddle/common/ddim.h"
+#include "paddle/fluid/pir/dialect/operator/ir/op_type.h"
+#include "paddle/pir/include/dialect/control_flow/ir/cf_op.h"
+
+// #include "paddle/cinn/frontend/group_pattern_util.h"
+
+namespace cinn {
+namespace hlir {
+namespace framework {
+namespace pir {
+namespace trivial_fusion_detail {
+
+struct TrivialOp {
+ public:
+  explicit TrivialOp(const ir::Expr& origin_func_body);
+
+  TrivialOp(const TrivialOp& trivial_op);
+
+  void _SetFuncBody(ir::Expr new_body);
+  ir::Expr* _GetFuncBodyPointer();
+
+  ir::Expr GetFuncBody() const;
+
+ private:
+  ir::Expr func_body;
+};
+
+struct ReduceOp {
+ public:
+  explicit ReduceOp(const ir::Expr& origin_func_body);
+  ReduceOp(const ReduceOp& reduce_op);
+
+  void _SetFuncBody(ir::Expr new_body);
+
+  ir::Expr GetFuncBody() const;
+
+  ir::Expr* _GetFuncBodyPointer();
+
+ private:
+  ir::Expr func_body;
+};
+
+using FusibleOp = std::variant<ReduceOp, TrivialOp>;
+
+ir::Expr _GetRootExpr(const FusibleOp& op);
+
+void _SetFuncBody(FusibleOp& op, ir::Expr new_body);
+ir::Expr GetComputeBody(const FusibleOp& op);
+
+ir::Tensor GetOutputTensor(const FusibleOp& op);
+
+ir::Expr _GetOriginalStoreValuePointer(const FusibleOp& op);
+
+std::vector<ir::Var> AppendBound(const std::vector<ir::Var> vars,
+                                 const ir::Expr& root);
+
+std::vector<ir::Var> GetOutputIters(const FusibleOp& op);
+
+std::vector<ir::Var> GetAllIterVars(const ReduceOp& op);
+
+std::vector<ir::Var> GetReduceIters(const ReduceOp& op);
+
+ir::Expr GetInitExpr(const ReduceOp& op);
+
+ir::Expr* _GetFuncBodyPointer(FusibleOp op);
+
+ir::Expr CopyReduceBody(const FusibleOp& downstream, const ReduceOp& upstream);
+
+ir::Expr CreateReduceExpr(
+    const std::vector<ir::Var>& output_iters,
+    const std::vector<ir::Var>& reduce_iters,
+    const ir::Expr& init_body,    // relay on output_iters
+    const ir::Expr& reduce_body,  // relay on output_iters + reduce_iters
+    const ir::Tensor& new_write_tensor,
+    const ir::Tensor& origin_write_tensor);
+
+ir::Expr CreateTrivialExpr(const std::vector<ir::Var>& output_iters,
+                           const ir::Expr& function_body,
+                           const ir::Tensor& new_write_tensor);
+ir::Expr CreateExprWithNewComputeBody(FusibleOp fusible_op,
+                                      ir::Expr new_compute_body);
+struct FusionNode {
+  FusibleOp fusible_op;
+  ::pir::Operation* expr_related_op;
+
+  std::unordered_map<FusionNode*, ::pir::Value> upstream;
+  std::unordered_map<FusionNode*, ::pir::Value> downstream;
+
+  explicit FusionNode(FusibleOp fusible_op);
+
+  static std::string GetTensorCounter();
+  void replace_topo_structure_of_fused_nodes(FusionNode* fused_up_node,
+                                             FusionNode* fused_down_node);
+
+  bool IsTrivial() const;
+};
+
+template <class DownStreamOp>
+DownStreamOp TrivalxOther_Fusion(TrivialOp upstream, DownStreamOp downstream) {
+  VLOG(4) << "Trivial x OtherFusion begin.";
+
+  const auto& replaced_tensor = GetOutputTensor(upstream);
+  VLOG(4) << "upstream is " << upstream.GetFuncBody();
+  VLOG(4) << "downstream is " << downstream.GetFuncBody();
+
+  DownStreamOp fused(ir::ir_utils::IRCopy(downstream.GetFuncBody()));
+  ir::Expr origin_compute_body = _GetOriginalStoreValuePointer(fused);
+  SequenceMutator(
+      ComposeUtils::GetEachTensorLoadExpr(origin_compute_body, replaced_tensor),
+      &origin_compute_body,
+      [&](const ir::Expr& downstream_load_expr, ir::Expr* downstream_body) {
+        ComposeUtils::ReplaceDownstreamLoadExprWithUpstreamComputeBody(
+            upstream, downstream_load_expr, downstream_body);
+      });
+
+  VLOG(4) << "After mutate, compute body: " << origin_compute_body;
+  VLOG(4) << "TTFusion end:\n" << fused.GetFuncBody();
+  return fused;
+}
+
+bool CheckAllLoopRangeEq(ReduceOp reduce_upper, TrivialOp trivial_down);
+
+std::vector<FusibleOp> TransformReduceLoopRange(const ReduceOp& upstream,
+                                                FusibleOp* downstream);
+
+FusibleOp TrivialFusion(FusionNode* upstream, FusionNode* downstream);
+
+FusibleOp SinkTrivialLoopAlign(TrivialOp trivial_op, ReduceOp reduce_op);
+
+std::vector<FusibleOp> ReduceTransformRecursive(FusibleOp root_op,
+                                                FusionNode* fusion_tree);
+std::vector<FusibleOp> ReduceTransform(FusionNode* downstream);
+
+FusibleOp CreateFusibleOp(ir::Expr compute_body, OpPatternKind op_pattern);
+
+struct FusionGraph {
+  explicit FusionGraph(const std::vector<::pir::Operation*>& ops,
+                       const std::vector<ir::Expr>& op_compute_bodies);
+
+  ~FusionGraph();
+
+  std::vector<ir::Expr> DoFusion();
+
+ private:
+  FusionNode* FindTrivialFusibleNode();
+
+  void DoTrivialFusion();
+
+  void ReduceLoopTranform();
+
+  std::vector<ir::Expr> GetExprResults();
+
+  void RemoveNode(FusionNode* node);
+
+  void AppendNode(FusionNode* node);
+
+  FusionNode* FindReduceUpstream(FusionNode* node);
+
+ private:
+  std::unordered_set<FusionNode*> all_fusion_nodes_;
+  std::vector<FusibleOp> fusion_results_;
+  std::unordered_set<FusionNode*> entrance_nodes_;
+  std::unordered_set<FusionNode*> exit_nodes_;
+
+  // std::unordered_map<::pir::Value, ShardableAxes> shardable_axes_;
+};
+
+}  // namespace trivial_fusion_detail
+
+std::vector<ir::Expr> OperationFusion(
+    const std::vector<::pir::Operation*>& ops,
+    const std::vector<ir::Expr>& op_compute_bodies);
+
+}  // namespace pir
+}  // namespace framework
+}  // namespace hlir
+}  // namespace cinn
